requirements.txt
streamlit
reportlab
apscheduler
python-dotenv
openai